Extract rrdata manipulation from Google provider
